refined the topo-stac-creation processes and moved its functions into seperate files
